optimise, improve
v0.1

+++ commands/removewarn.js
@@ -27,6 +27,15 @@ module.exports = {
                         .setName('confirmation')
                         .setDescription('Type in CONFIRM to confirm you want to remove EVERY WARNING ON THIS SERVER!')
                         .setRequired(true)))
+        .addSubcommand(subcommand =>
+                            subcommand
+                                .setName('clear_user')
+                                .setDescription('Clear every warning for a user.')
+                                .addUserOption(option =>
+                                    option
+                                        .setName('user')
+                                        .setDescription('User to removbe warns from.')
+                                        .setRequired(true)))
         .setDefaultMemberPermissions(PermissionFlagsBits.KickMembers),
 	async execute(interaction) {
         if (!interaction.member.permissions.has(PermissionsBitField.Flags.KickMembers)) {
@@ -46,7 +55,17 @@ module.exports = {
             } else {
                 await interaction.reply({ content : 'Please confirm you want to remove every message', ephemeral: true });
             }
-            } else {
+            } else if (interaction.options.getSubcommand() === "clear_user") {
+                    try {
+                        await Warn.deleteMany({ guidldId: interaction.guild.id, userId: interaction.options.getUser('user').id });
+                        await interaction.reply({ content: `All warns for ${interaction.options.getUser('user').username} have been removed.`, ephemeral: true});
+                    } catch (error) {
+                        await interaction.reply(error.message);
+                        console.warn(error)
+                        console.warn(`removewarn command failed.`)
+                    }  
+
+                } else {
                 const id = interaction.options.getString("id")
                 try {
                 await Warn.deleteOne({ _id: id });
